@@ -28,13 +28,13 @@
  */
 PandarGeneral::PandarGeneral(
     std::string device_ip, uint16_t lidar_port, uint16_t gps_port,
-    boost::function<void(boost::shared_ptr<PPointCloud>, double, hesai_lidar::PandarScanPtr)> pcl_callback,
+    boost::function<void(boost::shared_ptr<PPointCloud>, double, hesai_msgs::PandarScanPtr)> pcl_callback,
     boost::function<void(double)> gps_callback, uint16_t start_angle, int tz,
-    int pcl_type, std::string frame_id, std::string timestampType) {
+    int pcl_type, std::string lidar_type, std::string frame_id, std::string timestampType) {
       // LOG_FUNC();
   internal_ =
       new PandarGeneral_Internal(device_ip, lidar_port, gps_port, pcl_callback,
-                             gps_callback, start_angle, tz, pcl_type, frame_id, timestampType);
+                             gps_callback, start_angle, tz, pcl_type, lidar_type, frame_id, timestampType);
 }
 
 /**
@@ -47,10 +47,10 @@ PandarGeneral::PandarGeneral(
  */
 PandarGeneral::PandarGeneral(
     std::string pcap_path, \
-    boost::function<void(boost::shared_ptr<PPointCloud>, double, hesai_lidar::PandarScanPtr)> pcl_callback,\
-    uint16_t start_angle, int tz, int pcl_type, std::string frame_id, std::string timestampType) {
+    boost::function<void(boost::shared_ptr<PPointCloud>, double, hesai_msgs::PandarScanPtr)> pcl_callback,\
+    uint16_t start_angle, int tz, int pcl_type, std::string lidar_type, std::string frame_id, std::string timestampType) {
   internal_ = new PandarGeneral_Internal(pcap_path, pcl_callback, start_angle, \
-      tz, pcl_type, frame_id, timestampType);
+      tz, pcl_type, lidar_type, frame_id, timestampType);
 }
 
 /**
@@ -84,7 +84,7 @@ int PandarGeneral::Start() { return internal_->Start(); }
  */
 void PandarGeneral::Stop() { internal_->Stop(); }
 
-void PandarGeneral::PushScanPacket(hesai_lidar::PandarScanPtr scan) {
+void PandarGeneral::PushScanPacket(hesai_msgs::PandarScanPtr scan) {
   if (internal_) {
     internal_->PushScanPacket(scan);
   }
