@@ -18,20 +18,22 @@
 #include "src/tcp_command_client.h"
 #include "yaml-cpp/yaml.h"
 #include "log.h"
+#include "version.h"
 
 #define PANDARGENERALSDK_TCP_COMMAND_PORT (9347)
 
 PandarGeneralSDK::PandarGeneralSDK(
     std::string device_ip, const uint16_t lidar_port, const uint16_t gps_port,
-    boost::function<void(boost::shared_ptr<PPointCloud>, double, hesai_lidar::PandarScanPtr)>
+    boost::function<void(boost::shared_ptr<PPointCloud>, double, hesai_msgs::PandarScanPtr)>
         pcl_callback,
     boost::function<void(double)> gps_callback, uint16_t start_angle,
-    int tz, int pcl_type, std::string frame_id, std::string timestampType) {
+    int tz, int pcl_type, std::string lidar_type, std::string frame_id, std::string timestampType) {
+  printVersion();
   pandarGeneral_ = NULL;
   // LOG_FUNC();
 
   pandarGeneral_ = new PandarGeneral(device_ip, lidar_port,
-            gps_port, pcl_callback, gps_callback, start_angle, tz, pcl_type, frame_id, timestampType);
+            gps_port, pcl_callback, gps_callback, start_angle, tz, pcl_type, lidar_type, frame_id, timestampType);
 
   tcp_command_client_ =
       TcpCommandClientNew(device_ip.c_str(), PANDARGENERALSDK_TCP_COMMAND_PORT);
@@ -45,12 +47,13 @@ PandarGeneralSDK::PandarGeneralSDK(
 
 PandarGeneralSDK::PandarGeneralSDK(\
     std::string pcap_path, \
-    boost::function<void(boost::shared_ptr<PPointCloud>, double, hesai_lidar::PandarScanPtr)> pcl_callback, \
-    uint16_t start_angle, int tz, int pcl_type, std::string frame_id, std::string timestampType) {
+    boost::function<void(boost::shared_ptr<PPointCloud>, double, hesai_msgs::PandarScanPtr)> pcl_callback, \
+    uint16_t start_angle, int tz, int pcl_type, std::string lidar_type, std::string frame_id, std::string timestampType) {
+  printVersion();
   pandarGeneral_ = NULL;
 
   pandarGeneral_ = new PandarGeneral(pcap_path, pcl_callback, start_angle, \
-      tz, pcl_type, frame_id, timestampType);
+      tz, pcl_type, lidar_type, frame_id, timestampType);
 
   get_calibration_thr_ = NULL;
   tcp_command_client_ = NULL;
@@ -145,7 +148,7 @@ void PandarGeneralSDK::GetCalibrationFromDevice() {
   }
 }
 
-void PandarGeneralSDK::PushScanPacket(hesai_lidar::PandarScanPtr scan) {
+void PandarGeneralSDK::PushScanPacket(hesai_msgs::PandarScanPtr scan) {
   if (pandarGeneral_) {
     pandarGeneral_->PushScanPacket(scan);
   }
