@@ -78,13 +78,16 @@ static const float pandar20_horizatal_azimuth_offset_map[] = {
   -1.042f, -1.042f, -1.042f, -1.042f
 };
 
-static std::vector<std::vector<PPoint> > PointCloudList(128);
+static std::vector<std::vector<PPoint> > PointCloudList(MAX_LASER_NUM);
+// static std::vector<PPoint> PointCloud;
+static std::vector<PPoint> PointCloud(MAX_POINT_CLOUD_NUM);
+static int iPointCloudIndex = 0;
 
 PandarGeneral_Internal::PandarGeneral_Internal(
     std::string device_ip, uint16_t lidar_port, uint16_t gps_port,
-    boost::function<void(boost::shared_ptr<PPointCloud>, double, hesai_lidar::PandarScanPtr)> pcl_callback,
+    boost::function<void(boost::shared_ptr<PPointCloud>, double, hesai_msgs::PandarScanPtr)> pcl_callback,
     boost::function<void(double)> gps_callback, uint16_t start_angle, int tz,
-    int pcl_type, std::string frame_id, std::string timestampType) {
+    int pcl_type, std::string lidar_type, std::string frame_id, std::string timestampType) {
       // LOG_FUNC();
   pthread_mutex_init(&lidar_lock_, NULL);
   sem_init(&lidar_sem_, 0, 0);
@@ -101,6 +104,7 @@ PandarGeneral_Internal::PandarGeneral_Internal(
   pcl_callback_ = pcl_callback;
   gps_callback_ = gps_callback;
   last_azimuth_ = 0;
+  m_sLidarType = lidar_type;
   frame_id_ = frame_id;
   tz_second_ = tz * 3600;
   pcl_type_ = pcl_type;
@@ -113,9 +117,9 @@ PandarGeneral_Internal::PandarGeneral_Internal(
 }
 
 PandarGeneral_Internal::PandarGeneral_Internal(std::string pcap_path, \
-    boost::function<void(boost::shared_ptr<PPointCloud>, double, hesai_lidar::PandarScanPtr)> \
+    boost::function<void(boost::shared_ptr<PPointCloud>, double, hesai_msgs::PandarScanPtr)> \
     pcl_callback, uint16_t start_angle, int tz, int pcl_type, \
-    std::string frame_id, std::string timestampType) {
+    std::string lidar_type, std::string frame_id, std::string timestampType) {
   pthread_mutex_init(&lidar_lock_, NULL);
   sem_init(&lidar_sem_, 0, 0);
 
@@ -125,19 +129,19 @@ PandarGeneral_Internal::PandarGeneral_Internal(std::string pcap_path, \
   enable_lidar_recv_thr_ = false;
   enable_lidar_process_thr_ = false;
 
-  pcap_reader_ = new PcapReader(pcap_path,frame_id);
+  pcap_reader_ = new PcapReader(pcap_path,lidar_type);
 
   start_angle_ = start_angle;
   pcl_callback_ = pcl_callback;
   gps_callback_ = NULL;
   last_azimuth_ = 0;
+  m_sLidarType = lidar_type;
   frame_id_ = frame_id;
   tz_second_ = tz * 3600;
   pcl_type_ = pcl_type;
   connect_lidar_ = false;
   m_sTimestampType = timestampType;
   m_dPktTimestamp = 0.0f;
-
   Init();
 }
 
@@ -157,6 +161,17 @@ void PandarGeneral_Internal::Init() {
     float rotation = degreeToRadian(0.01 * static_cast<double>(rotIndex));
     cos_lookup_table_[rotIndex] = cosf(rotation);
     sin_lookup_table_[rotIndex] = sinf(rotation);
+  }
+  m_sin_azimuth_map_.resize(MAX_AZIMUTH_DEGREE_NUM);
+  m_cos_azimuth_map_.resize(MAX_AZIMUTH_DEGREE_NUM);
+  for(int i = 0; i < MAX_AZIMUTH_DEGREE_NUM; ++i) {
+    m_sin_azimuth_map_[i] = sinf(i * M_PI / 18000);
+    m_cos_azimuth_map_[i] = cosf(i * M_PI / 18000);
+  }
+  if (pcl_type_) {
+    for (int i = 0; i < MAX_LASER_NUM; i++) {
+      PointCloudList[i].reserve(MAX_POINT_CLOUD_NUM_PER_CHANNEL);
+    }
   }
 
   //laser40
@@ -433,46 +448,70 @@ void PandarGeneral_Internal::Init() {
   laserQTOffset_[62] = 10.0f + 134.39f;
   laserQTOffset_[63] = 10.0f + 136.45f;
 
-  if (frame_id_ == "Pandar40P" || frame_id_ == "Pandar40M") {
+  if (m_sLidarType == "Pandar40P" || m_sLidarType == "Pandar40M") {
+    m_sin_elevation_map_.resize(LASER_COUNT);
+    m_cos_elevation_map_.resize(LASER_COUNT);
     for (int i = 0; i < LASER_COUNT; i++) {
+      m_sin_elevation_map_[i] = sinf(degreeToRadian(pandar40p_elev_angle_map[i]));
+      m_cos_elevation_map_[i] = cosf(degreeToRadian(pandar40p_elev_angle_map[i]));
       General_elev_angle_map_[i] = pandar40p_elev_angle_map[i];
       General_horizatal_azimuth_offset_map_[i] = \
           pandar40p_horizatal_azimuth_offset_map[i];
     }
   }
 
-  if (frame_id_ == "Pandar64") {
+  if (m_sLidarType == "Pandar64") {
+    m_sin_elevation_map_.resize(HS_LIDAR_L64_UNIT_NUM);
+    m_cos_elevation_map_.resize(HS_LIDAR_L64_UNIT_NUM);
     for (int i = 0; i < HS_LIDAR_L64_UNIT_NUM; i++) {
+      m_sin_elevation_map_[i] = sinf(degreeToRadian(pandarGeneral_elev_angle_map[i]));
+      m_cos_elevation_map_[i] = cosf(degreeToRadian(pandarGeneral_elev_angle_map[i]));
       General_elev_angle_map_[i] = pandarGeneral_elev_angle_map[i];
       General_horizatal_azimuth_offset_map_[i] = \
           pandarGeneral_horizatal_azimuth_offset_map[i];
     }
   }
 
-  if (frame_id_ == "Pandar20A" || frame_id_ == "Pandar20B") {
+  if (m_sLidarType == "Pandar20A" || m_sLidarType == "Pandar20B") {
+    m_sin_elevation_map_.resize(HS_LIDAR_L20_UNIT_NUM);
+    m_cos_elevation_map_.resize(HS_LIDAR_L20_UNIT_NUM);
     for (int i = 0; i < HS_LIDAR_L20_UNIT_NUM; i++) {
+      m_sin_elevation_map_[i] = sinf(degreeToRadian(pandar20_elev_angle_map[i]));
+      m_cos_elevation_map_[i] = cosf(degreeToRadian(pandar20_elev_angle_map[i]));
       General_elev_angle_map_[i] = pandar20_elev_angle_map[i];
       General_horizatal_azimuth_offset_map_[i] = pandar20_horizatal_azimuth_offset_map[i];
     }
   }
 
-  if (frame_id_ == "PandarQT") {
+  if (m_sLidarType == "PandarQT") {
+    m_sin_elevation_map_.resize(HS_LIDAR_QT_UNIT_NUM);
+    m_cos_elevation_map_.resize(HS_LIDAR_QT_UNIT_NUM);
     for (int i = 0; i < HS_LIDAR_QT_UNIT_NUM; i++) {
+      m_sin_elevation_map_[i] = sinf(degreeToRadian(pandarQT_elev_angle_map[i]));
+      m_cos_elevation_map_[i] = cosf(degreeToRadian(pandarQT_elev_angle_map[i]));
       General_elev_angle_map_[i] = pandarQT_elev_angle_map[i];
       General_horizatal_azimuth_offset_map_[i] = pandarQT_horizatal_azimuth_offset_map[i];
     }
   }
 
-  if (frame_id_ == "PandarXT-32") {
+  if (m_sLidarType == "PandarXT-32") {
+    m_sin_elevation_map_.resize(HS_LIDAR_XT_UNIT_NUM);
+    m_cos_elevation_map_.resize(HS_LIDAR_XT_UNIT_NUM);
     for (int i = 0; i < HS_LIDAR_XT_UNIT_NUM; i++) {
+      m_sin_elevation_map_[i] = sinf(degreeToRadian(pandarXT_elev_angle_map[i]));
+      m_cos_elevation_map_[i] = cosf(degreeToRadian(pandarXT_elev_angle_map[i]));
       General_elev_angle_map_[i] = pandarXT_elev_angle_map[i];
       General_horizatal_azimuth_offset_map_[i] = pandarXT_horizatal_azimuth_offset_map[i];
       laserXTOffset_[i] = laserXTOffset[i];
     }
   }
 
-  if (frame_id_ == "PandarXT-16") {
+  if (m_sLidarType == "PandarXT-16") {
+    m_sin_elevation_map_.resize(HS_LIDAR_XT16_UNIT_NUM);
+    m_cos_elevation_map_.resize(HS_LIDAR_XT16_UNIT_NUM);
     for (int i = 0; i < HS_LIDAR_XT16_UNIT_NUM; i++) {
+      m_sin_elevation_map_[i] = sinf(degreeToRadian(pandarXT_elev_angle_map[i*2]));
+      m_cos_elevation_map_[i] = cosf(degreeToRadian(pandarXT_elev_angle_map[i*2]));
       General_elev_angle_map_[i] = pandarXT_elev_angle_map[i*2];
       General_horizatal_azimuth_offset_map_[i] = pandarXT_horizatal_azimuth_offset_map[i*2];
       laserXTOffset_[i] = laserXTOffset[i*2];
@@ -483,6 +522,7 @@ void PandarGeneral_Internal::Init() {
     blockXTOffsetSingle_[i] = blockXTOffsetSingle[i];
     blockXTOffsetDual_[i] = blockXTOffsetDual[i];
   }
+  SetEnvironmentVariableTZ();
 }
 
 /**
@@ -529,10 +569,13 @@ int PandarGeneral_Internal::LoadCorrectionFile(std::string correction_content) {
     elev_angle[lineId - 1] = elev;
     azimuthOffset[lineId - 1] = azimuth;
   }
-
+  m_sin_elevation_map_.resize(lineCounter);
+  m_cos_elevation_map_.resize(lineCounter);
   for (int i = 0; i < lineCounter; ++i) {
     /* for all the laser offset */
     General_elev_angle_map_[i] = elev_angle[i];
+    m_sin_elevation_map_[i] = sinf(degreeToRadian(General_elev_angle_map_[i]));
+    m_cos_elevation_map_[i] = cosf(degreeToRadian(General_elev_angle_map_[i]));
     General_horizatal_azimuth_offset_map_[i] = azimuthOffset[i];
   }
 
@@ -626,8 +669,8 @@ void PandarGeneral_Internal::ProcessLiarPacket() {
   int ret = 0;
 
   boost::shared_ptr<PPointCloud> outMsg(new PPointCloud());
-  hesai_lidar::PandarScanPtr scan(new hesai_lidar::PandarScan);
-  hesai_lidar::PandarPacket rawpacket;
+  hesai_msgs::PandarScanPtr scan(new hesai_msgs::PandarScan);
+  hesai_msgs::PandarPacket rawpacket;
 
 
   while (enable_lidar_process_thr_) {
@@ -644,6 +687,11 @@ void PandarGeneral_Internal::ProcessLiarPacket() {
     lidar_packets_.pop_front();
     pthread_mutex_unlock(&lidar_lock_);
     memcpy(&rawpacket, &packet, 1512);
+//    rawpacket.stamp.sec = floor(packet.stamp);
+//    rawpacket.stamp.nsec = (packet.stamp - floor(packet.stamp))*1000000000;
+//    rawpacket.size = packet.size;
+//    rawpacket.data.resize(packet.size);
+//    memcpy(&rawpacket.data[0], &packet.data[0], packet.size);
     m_dPktTimestamp = packet.stamp;
     // printf("##m_dPktTimestamp: %lf\n", m_dPktTimestamp);
 
@@ -671,10 +719,9 @@ void PandarGeneral_Internal::ProcessLiarPacket() {
                start_angle_ <= pkt.blocks[i].azimuth) ||
               (last_azimuth_ < start_angle_ &&
                start_angle_ <= pkt.blocks[i].azimuth)) {
-            if (pcl_callback_ && (outMsg->points.size() > 0 || PointCloudList[0].size() > 0)) {
+            if (pcl_callback_ && (iPointCloudIndex > 0 || PointCloudList[0].size() > 0)) {
               EmitBackMessege(LASER_COUNT, outMsg, scan);
               scan->packets.clear();
-              outMsg.reset(new PPointCloud());
             }
           }
         }
@@ -707,10 +754,9 @@ void PandarGeneral_Internal::ProcessLiarPacket() {
                start_angle_ <= pkt.blocks[i].azimuth) ||
               (last_azimuth_ < start_angle_ &&
                start_angle_ <= pkt.blocks[i].azimuth)) {
-            if (pcl_callback_ && (outMsg->points.size() > 0 || PointCloudList[0].size() > 0)) {
+            if (pcl_callback_ && (iPointCloudIndex > 0 || PointCloudList[0].size() > 0)) {
               EmitBackMessege(pkt.header.chLaserNumber, outMsg, scan);
               scan->packets.clear();
-              outMsg.reset(new PPointCloud());
             }
           }
         } else {
@@ -743,10 +789,9 @@ void PandarGeneral_Internal::ProcessLiarPacket() {
                start_angle_ <= pkt.blocks[i].azimuth) ||
               (last_azimuth_ < start_angle_ &&
                start_angle_ <= pkt.blocks[i].azimuth)) {
-            if (pcl_callback_ && (outMsg->points.size() > 0 || PointCloudList[0].size() > 0)) {
+            if (pcl_callback_ && (iPointCloudIndex > 0 || PointCloudList[0].size() > 0)) {
               EmitBackMessege(pkt.header.chLaserNumber, outMsg, scan);
               scan->packets.clear();
-              outMsg.reset(new PPointCloud());
             }
           }
         } else {
@@ -778,10 +823,9 @@ void PandarGeneral_Internal::ProcessLiarPacket() {
                start_angle_ <= pkt.blocks[i].azimuth) ||
               (last_azimuth_ < start_angle_ &&
                start_angle_ <= pkt.blocks[i].azimuth)) {
-            if (pcl_callback_ && (outMsg->points.size() > 0 || PointCloudList[0].size() > 0)) {
+            if (pcl_callback_ && (iPointCloudIndex > 0 || PointCloudList[0].size() > 0)) {
               EmitBackMessege(pkt.header.chLaserNumber, outMsg, scan);
               scan->packets.clear();
-              outMsg.reset(new PPointCloud());
             }
           }
         } else {
@@ -791,8 +835,8 @@ void PandarGeneral_Internal::ProcessLiarPacket() {
         last_azimuth_ = pkt.blocks[i].azimuth;
       }
     } 
-    else if((packet.size == HS_LIDAR_XT_PACKET_SIZE && (frame_id_ == "XT" || frame_id_ == "PandarXT-32")) || \
-        (packet.size == HS_LIDAR_XT16_PACKET_SIZE && (frame_id_ == "PandarXT-16"))) {
+    else if((packet.size == HS_LIDAR_XT_PACKET_SIZE && (m_sLidarType == "XT" || m_sLidarType == "PandarXT-32")) || \
+        (packet.size == HS_LIDAR_XT16_PACKET_SIZE && (m_sLidarType == "PandarXT-16"))) {
       HS_LIDAR_XT_Packet pkt;
       ret = ParseXTData(&pkt, packet.data, packet.size);
       if (ret != 0) {
@@ -814,10 +858,9 @@ void PandarGeneral_Internal::ProcessLiarPacket() {
                start_angle_ <= pkt.blocks[i].azimuth) ||
               (last_azimuth_ < start_angle_ &&
                start_angle_ <= pkt.blocks[i].azimuth)) {
-            if (pcl_callback_ && (outMsg->points.size() > 0 || PointCloudList[0].size() > 0)) {
+            if (pcl_callback_ && (iPointCloudIndex > 0 || PointCloudList[0].size() > 0)) {
               EmitBackMessege(pkt.header.chLaserNumber, outMsg, scan);
               scan->packets.clear();
-              outMsg.reset(new PPointCloud());
             }
           }
         } else {
@@ -843,7 +886,7 @@ void PandarGeneral_Internal::PushLiDARData(PandarPacket packet) {
 }
 
 void PandarGeneral_Internal::ProcessGps(const PandarGPS &gpsMsg) {
-  struct tm t;
+  struct tm t = {0};
   t.tm_sec = gpsMsg.second;
   t.tm_min = gpsMsg.minute;
 
@@ -1260,17 +1303,15 @@ void PandarGeneral_Internal::CalcPointXYZIT(Pandar40PPacket *pkt, int blockid,
       continue;
     }
 
-    double xyDistance = unit.distance * \
-        cosf(degreeToRadian(General_elev_angle_map_[i]));
-
-    point.x = static_cast<float>(xyDistance * \
-        sinf(degreeToRadian(General_horizatal_azimuth_offset_map_[i] + \
-        (static_cast<double>(block->azimuth)) / 100.0)));
-    point.y = static_cast<float>(xyDistance * \
-        cosf(degreeToRadian(General_horizatal_azimuth_offset_map_[i] + \
-        (static_cast<double>(block->azimuth)) / 100.0)));
-    point.z = static_cast<float>(unit.distance * \
-        sinf(degreeToRadian(General_elev_angle_map_[i])));
+    int azimuth = static_cast<int>(General_horizatal_azimuth_offset_map_[i] * 100 + block->azimuth);
+    if(azimuth < 0)
+      azimuth += 36000;
+    if(azimuth > 36000)
+      azimuth -= 36000;
+    float xyDistance = unit.distance * m_cos_elevation_map_[i];
+    point.x = static_cast<float>(xyDistance * m_sin_azimuth_map_[azimuth]);
+    point.y = static_cast<float>(xyDistance * m_cos_azimuth_map_[azimuth]);
+    point.z = static_cast<float>(unit.distance * m_sin_elevation_map_[i]);
 
     point.intensity = unit.intensity;
 
@@ -1298,7 +1339,8 @@ void PandarGeneral_Internal::CalcPointXYZIT(Pandar40PPacket *pkt, int blockid,
     if (pcl_type_) {
       PointCloudList[i].push_back(point);
     } else {
-      cld->push_back(point);
+      PointCloud[iPointCloudIndex] = point;
+      iPointCloudIndex++;
     }
   }
 }
@@ -1307,7 +1349,7 @@ void PandarGeneral_Internal::CalcL64PointXYZIT(HS_LIDAR_L64_Packet *pkt, int blo
     char chLaserNumber, boost::shared_ptr<PPointCloud> cld) {
   HS_LIDAR_L64_Block *block = &pkt->blocks[blockid];
 
-  struct tm tTm;
+  struct tm tTm = {0};
   // UTC's year only include 0 - 99 year , which indicate 2000 to 2099.
   // and mktime's year start from 1900 which is 0. so we need add 100 year.
   tTm.tm_year = pkt->addtime[0] + 100;
@@ -1338,18 +1380,15 @@ void PandarGeneral_Internal::CalcL64PointXYZIT(HS_LIDAR_L64_Packet *pkt, int blo
       continue;
     }
 
-    double xyDistance = \
-        unit.distance * cosf(degreeToRadian(General_elev_angle_map_[i]));
-    point.x = static_cast<float>(
-        xyDistance *
-        sinf(degreeToRadian(General_horizatal_azimuth_offset_map_[i] +
-                            (static_cast<double>(block->azimuth)) / 100.0)));
-    point.y = static_cast<float>(
-        xyDistance *
-        cosf(degreeToRadian(General_horizatal_azimuth_offset_map_[i] +
-                            (static_cast<double>(block->azimuth)) / 100.0)));
-    point.z = static_cast<float>(unit.distance *
-                                 sinf(degreeToRadian(General_elev_angle_map_[i])));
+    int azimuth = static_cast<int>(General_horizatal_azimuth_offset_map_[i] * 100 + block->azimuth);
+    if(azimuth < 0)
+      azimuth += 36000;
+    if(azimuth > 36000)
+      azimuth -= 36000;
+    float xyDistance = unit.distance * m_cos_elevation_map_[i];
+    point.x = static_cast<float>(xyDistance * m_sin_azimuth_map_[azimuth]);
+    point.y = static_cast<float>(xyDistance * m_cos_azimuth_map_[azimuth]);
+    point.z = static_cast<float>(unit.distance * m_sin_elevation_map_[i]);
 
     point.intensity = unit.intensity;
 
@@ -1378,7 +1417,8 @@ void PandarGeneral_Internal::CalcL64PointXYZIT(HS_LIDAR_L64_Packet *pkt, int blo
     if (pcl_type_) {
       PointCloudList[i].push_back(point);
     } else {
-      cld->push_back(point);
+      PointCloud[iPointCloudIndex] = point;
+      iPointCloudIndex++;
     }
   }
 }
@@ -1387,7 +1427,7 @@ void PandarGeneral_Internal::CalcL20PointXYZIT(HS_LIDAR_L20_Packet *pkt, int blo
     char chLaserNumber, boost::shared_ptr<PPointCloud> cld) {
   HS_LIDAR_L20_Block *block = &pkt->blocks[blockid];
 
-  struct tm tTm;
+  struct tm tTm = {0};
   // UTC's year only include 0 - 99 year , which indicate 2000 to 2099.
   // and mktime's year start from 1900 which is 0. so we need add 100 year.
   tTm.tm_year = pkt->addtime[0] + 100;
@@ -1418,16 +1458,15 @@ void PandarGeneral_Internal::CalcL20PointXYZIT(HS_LIDAR_L20_Packet *pkt, int blo
       continue;
     }
 
-    double xyDistance = unit.distance * cosf(degreeToRadian(General_elev_angle_map_[i]));
-
-    point.x = static_cast<float>(xyDistance * \
-        sinf(degreeToRadian(General_horizatal_azimuth_offset_map_[i] + \
-        (static_cast<double>(block->azimuth)) / 100.0)));
-    point.y = static_cast<float>(xyDistance * \
-        cosf(degreeToRadian(General_horizatal_azimuth_offset_map_[i] + \
-        (static_cast<double>(block->azimuth)) / 100.0)));
-    point.z = static_cast<float>(unit.distance * \
-        sinf(degreeToRadian(General_elev_angle_map_[i])));
+    int azimuth = static_cast<int>(General_horizatal_azimuth_offset_map_[i] * 100 + block->azimuth);
+    if(azimuth < 0)
+      azimuth += 36000;
+    if(azimuth > 36000)
+      azimuth -= 36000;
+    float xyDistance = unit.distance * m_cos_elevation_map_[i];
+    point.x = static_cast<float>(xyDistance * m_sin_azimuth_map_[azimuth]);
+    point.y = static_cast<float>(xyDistance * m_cos_azimuth_map_[azimuth]);
+    point.z = static_cast<float>(unit.distance * m_sin_elevation_map_[i]);
 
     point.intensity = unit.intensity;
 
@@ -1440,21 +1479,21 @@ void PandarGeneral_Internal::CalcL20PointXYZIT(HS_LIDAR_L20_Packet *pkt, int blo
 
       if (pkt->echo == 0x39) {
         // dual return, block 0&1 (2&3 , 4*5 ...)'s timestamp is the same.
-        if (strcmp(frame_id_.c_str(), "Pandar20A") == 0) {
+        if (strcmp(m_sLidarType.c_str(), "Pandar20A") == 0) {
           point.timestamp = point.timestamp - \
               (static_cast<double>(block20OffsetDual_[blockid] + \
               laser20AOffset_[i]) / 1000000.0f);
-        } else if (strcmp(frame_id_.c_str(), "Pandar20B") == 0) {
+        } else if (strcmp(m_sLidarType.c_str(), "Pandar20B") == 0) {
           point.timestamp = point.timestamp - \
               (static_cast<double>(block20OffsetDual_[blockid] + \
               laser20BOffset_[i]) / 1000000.0f);
         }
       } else {
-        if (strcmp(frame_id_.c_str(), "Pandar20A") == 0) {
+        if (strcmp(m_sLidarType.c_str(), "Pandar20A") == 0) {
           point.timestamp = point.timestamp - \
               (static_cast<double>(block20OffsetSingle_[blockid] + \
               laser20AOffset_[i]) / 1000000.0f);
-        } else if (strcmp(frame_id_.c_str(), "Pandar20B") == 0) {
+        } else if (strcmp(m_sLidarType.c_str(), "Pandar20B") == 0) {
           point.timestamp = point.timestamp - \
               (static_cast<double>(block20OffsetSingle_[blockid] + \
               laser20BOffset_[i]) / 1000000.0f);
@@ -1467,7 +1506,8 @@ void PandarGeneral_Internal::CalcL20PointXYZIT(HS_LIDAR_L20_Packet *pkt, int blo
     if (pcl_type_) {
       PointCloudList[i].push_back(point);
     } else {
-      cld->push_back(point);
+      PointCloud[iPointCloudIndex] = point;
+      iPointCloudIndex++;
     }
   }
 }
@@ -1477,7 +1517,7 @@ void PandarGeneral_Internal::CalcQTPointXYZIT(HS_LIDAR_QT_Packet *pkt, int block
     char chLaserNumber, boost::shared_ptr<PPointCloud> cld) {
   HS_LIDAR_QT_Block *block = &pkt->blocks[blockid];
 
-  struct tm tTm;
+  struct tm tTm = {0};
   tTm.tm_year = pkt->addtime[0] + 100;
 
   // in case of time error
@@ -1506,16 +1546,15 @@ void PandarGeneral_Internal::CalcQTPointXYZIT(HS_LIDAR_QT_Packet *pkt, int block
       continue;
     }
 
-    double xyDistance = unit.distance * cosf(degreeToRadian(General_elev_angle_map_[i]));
-
-    point.x = static_cast<float>(xyDistance * \
-        sinf(degreeToRadian(General_horizatal_azimuth_offset_map_[i] + \
-        (static_cast<double>(block->azimuth)) / 100.0)));
-    point.y = static_cast<float>(xyDistance * \
-        cosf(degreeToRadian(General_horizatal_azimuth_offset_map_[i] + \
-        (static_cast<double>(block->azimuth)) / 100.0)));
-    point.z = static_cast<float>(unit.distance * \
-        sinf(degreeToRadian(General_elev_angle_map_[i])));
+    int azimuth = static_cast<int>(General_horizatal_azimuth_offset_map_[i] * 100 + block->azimuth);
+    if(azimuth < 0)
+      azimuth += 36000;
+    if(azimuth > 36000)
+      azimuth -= 36000;
+    float xyDistance = unit.distance * m_cos_elevation_map_[i];
+    point.x = static_cast<float>(xyDistance * m_sin_azimuth_map_[azimuth]);
+    point.y = static_cast<float>(xyDistance * m_cos_azimuth_map_[azimuth]);
+    point.z = static_cast<float>(unit.distance * m_sin_elevation_map_[i]);
 
     point.intensity = unit.intensity;
 
@@ -1544,7 +1583,8 @@ void PandarGeneral_Internal::CalcQTPointXYZIT(HS_LIDAR_QT_Packet *pkt, int block
     if (pcl_type_)
       PointCloudList[i].push_back(point);
     else
-      cld->push_back(point);
+      PointCloud[iPointCloudIndex] = point;
+      iPointCloudIndex++;
   }
 }
 
@@ -1552,7 +1592,7 @@ void PandarGeneral_Internal::CalcXTPointXYZIT(HS_LIDAR_XT_Packet *pkt, int block
     char chLaserNumber, boost::shared_ptr<PPointCloud> cld) {
   HS_LIDAR_XT_Block *block = &pkt->blocks[blockid];
 
-  struct tm tTm;
+  struct tm tTm = {0};
   tTm.tm_year = pkt->addtime[0] + 100;
 
   // in case of time error
@@ -1581,16 +1621,15 @@ void PandarGeneral_Internal::CalcXTPointXYZIT(HS_LIDAR_XT_Packet *pkt, int block
       continue;
     }
 
-    double xyDistance = unit.distance * cosf(degreeToRadian(General_elev_angle_map_[i]));
-
-    point.x = static_cast<float>(xyDistance * \
-        sinf(degreeToRadian(General_horizatal_azimuth_offset_map_[i] + \
-        (static_cast<double>(block->azimuth)) / 100.0)));
-    point.y = static_cast<float>(xyDistance * \
-        cosf(degreeToRadian(General_horizatal_azimuth_offset_map_[i] + \
-        (static_cast<double>(block->azimuth)) / 100.0)));
-    point.z = static_cast<float>(unit.distance * \
-        sinf(degreeToRadian(General_elev_angle_map_[i])));
+    int azimuth = static_cast<int>(General_horizatal_azimuth_offset_map_[i] * 100 + block->azimuth);
+    if(azimuth < 0)
+      azimuth += 36000;
+    if(azimuth > 36000)
+      azimuth -= 36000;
+    float xyDistance = unit.distance * m_cos_elevation_map_[i];
+    point.x = static_cast<float>(xyDistance * m_sin_azimuth_map_[azimuth]);
+    point.y = static_cast<float>(xyDistance * m_cos_azimuth_map_[azimuth]);
+    point.z = static_cast<float>(unit.distance * m_sin_elevation_map_[i]);
 
     point.intensity = unit.intensity;
 
@@ -1617,12 +1656,13 @@ void PandarGeneral_Internal::CalcXTPointXYZIT(HS_LIDAR_XT_Packet *pkt, int block
     if (pcl_type_) {
       PointCloudList[i].push_back(point);
     } else {
-      cld->push_back(point);
+      PointCloud[iPointCloudIndex] = point;
+      iPointCloudIndex++;
     }
   }
 }
 
-void PandarGeneral_Internal::EmitBackMessege(char chLaserNumber, boost::shared_ptr<PPointCloud> cld, hesai_lidar::PandarScanPtr scan) {
+void PandarGeneral_Internal::EmitBackMessege(char chLaserNumber, boost::shared_ptr<PPointCloud> cld, hesai_msgs::PandarScanPtr scan) {
   if (pcl_type_) {
     for (int i=0; i<chLaserNumber; i++) {
       for (int j=0; j<PointCloudList[i].size(); j++) {
@@ -1630,20 +1670,59 @@ void PandarGeneral_Internal::EmitBackMessege(char chLaserNumber, boost::shared_p
       }
     }
   }
+  else{
+    cld->points.assign(PointCloud.begin(),PointCloud.begin() + iPointCloudIndex);
+    cld->width = (uint32_t)cld->points.size();
+    cld->height = 1;
+    iPointCloudIndex = 0;
+  }
   pcl_callback_(cld, cld->points[0].timestamp, scan); // the timestamp from first point cloud of cld
-  //cld.reset(new PPointCloud());
   if (pcl_type_) {
-    for (int i=0; i<128; i++) {
+    for (int i=0; i<chLaserNumber; i++) {
       PointCloudList[i].clear();
+      PointCloudList[i].reserve(MAX_POINT_CLOUD_NUM_PER_CHANNEL);
+      cld->points.clear();
+      cld->width = (uint32_t)cld->points.size();
+      cld->height = 1;
     }
   }
 }
 
-void PandarGeneral_Internal::PushScanPacket(hesai_lidar::PandarScanPtr scan) {
+void PandarGeneral_Internal::PushScanPacket(hesai_msgs::PandarScanPtr scan) {
   for(int i = 0; i<scan->packets.size(); i++) {
     PandarPacket pkt;
-    memcpy(&pkt, &scan->packets[i], 1512);
+    pkt.stamp = scan->packets[i].stamp.sec + scan->packets[i].stamp.nsec / 1000000000.0;
+    pkt.size = scan->packets[i].size;
+    memcpy(&pkt.data[0], &(scan->packets[i].data[0]), scan->packets[i].size);
     PushLiDARData(pkt);
+  }
+}
+
+void PandarGeneral_Internal::SetEnvironmentVariableTZ(){
+  char *TZ; 
+  if((TZ = getenv("TZ"))){
+    printf("TZ=%s\n",TZ); 
+    return;
+  } 
+  unsigned int timezone = 0;
+  time_t t1, t2 ;
+  struct tm *tm_local, *tm_utc;
+  time(&t1);
+  t2 = t1;
+  tm_local = localtime(&t1);
+  t1 = mktime(tm_local) ;
+  tm_utc = gmtime(&t2);
+  t2 = mktime(tm_utc);
+  timezone = t2 >= t1 ? (t2 - t1) / 3600 : (t1 - t2) / 3600;
+  std::string data = "TZ=UTC" + std::to_string(timezone);
+  int len = data.length();
+  TZ = (char *)malloc((len + 1) * sizeof(char));
+  data.copy(TZ, len, 0); 
+  if(putenv(TZ) == 0){
+    printf("set environment %s\n", TZ);
+  }
+  else{
+    printf("set environment fail\n");
   }
 }
 
