@@ -32,8 +32,8 @@
 #include "src/pandarQT.h"
 #include "src/pandarXT.h"
 #include "src/pcap_reader.h"
-#include "hesai_lidar/PandarScan.h"
-#include "hesai_lidar/PandarPacket.h"
+#include "hesai_msgs/PandarScan.h"
+#include "hesai_msgs/PandarPacket.h"
 
 #define SOB_ANGLE_SIZE (4)
 #define RAW_MEASURE_SIZE (3)
@@ -136,6 +136,9 @@ HS_LIDAR_L64_7_BLOCK_PACKET_BODY_SIZE + HS_LIDAR_L64_PACKET_TAIL_WITHOUT_UDPSEQ_
 #define HesaiLidarSDK_DEFAULT_GPS_RECV_PORT 10110
 
 #define MAX_LASER_NUM (256)
+#define MAX_POINT_CLOUD_NUM (1000000)
+#define MAX_POINT_CLOUD_NUM_PER_CHANNEL (10000)
+#define MAX_AZIMUTH_DEGREE_NUM (36000)
 
 struct Pandar40PUnit_s {
   uint8_t intensity;
@@ -264,9 +267,9 @@ class PandarGeneral_Internal {
    */
   PandarGeneral_Internal(
       std::string device_ip, uint16_t lidar_port, uint16_t gps_port,
-      boost::function<void(boost::shared_ptr<PPointCloud>, double, hesai_lidar::PandarScanPtr)>
+      boost::function<void(boost::shared_ptr<PPointCloud>, double, hesai_msgs::PandarScanPtr)>
           pcl_callback, boost::function<void(double)> gps_callback, 
-      uint16_t start_angle, int tz, int pcl_type, std::string frame_id, std::string timestampType);
+      uint16_t start_angle, int tz, int pcl_type, std::string lidar_type, std::string frame_id, std::string timestampType);
 
   /**
    * @brief Constructor
@@ -279,9 +282,9 @@ class PandarGeneral_Internal {
    */
   PandarGeneral_Internal(
       std::string pcap_path, \
-      boost::function<void(boost::shared_ptr<PPointCloud>, double, hesai_lidar::PandarScanPtr)> \
+      boost::function<void(boost::shared_ptr<PPointCloud>, double, hesai_msgs::PandarScanPtr)> \
       pcl_callback, uint16_t start_angle, int tz, int pcl_type, \
-      std::string frame_id, std::string timestampType);// the default timestamp type is LiDAR time
+      std::string lidar_type, std::string frame_id, std::string timestampType);// the default timestamp type is LiDAR time
   ~PandarGeneral_Internal();
 
   /**
@@ -298,7 +301,7 @@ class PandarGeneral_Internal {
 
   int Start();
   void Stop();
-  void PushScanPacket(hesai_lidar::PandarScanPtr scan);
+  void PushScanPacket(hesai_msgs::PandarScanPtr scan);
 
  private:
   void Init();
@@ -325,7 +328,8 @@ class PandarGeneral_Internal {
                       boost::shared_ptr<PPointCloud> cld);
   void FillPacket(const uint8_t *buf, const int len, double timestamp);
 
-  void EmitBackMessege(char chLaserNumber, boost::shared_ptr<PPointCloud> cld, hesai_lidar::PandarScanPtr scan);
+  void EmitBackMessege(char chLaserNumber, boost::shared_ptr<PPointCloud> cld, hesai_msgs::PandarScanPtr scan);
+  void SetEnvironmentVariableTZ();
   pthread_mutex_t lidar_lock_;
   sem_t lidar_sem_;
   boost::thread *lidar_recv_thr_;
@@ -339,7 +343,7 @@ class PandarGeneral_Internal {
   std::list<struct PandarPacket_s> lidar_packets_;
 
   boost::shared_ptr<Input> input_;
-  boost::function<void(boost::shared_ptr<PPointCloud> cld, double timestamp, hesai_lidar::PandarScanPtr scan)>
+  boost::function<void(boost::shared_ptr<PPointCloud> cld, double timestamp, hesai_msgs::PandarScanPtr scan)>
       pcl_callback_;
   boost::function<void(double timestamp)> gps_callback_;
 
@@ -389,6 +393,12 @@ class PandarGeneral_Internal {
   int pcl_type_;
   PcapReader *pcap_reader_;
   bool connect_lidar_;
+  std::string m_sLidarType;
+  std::vector<float> m_sin_azimuth_map_;
+  std::vector<float> m_cos_azimuth_map_;
+  std::vector<float> m_sin_elevation_map_;
+  std::vector<float> m_cos_elevation_map_;
+
 };
 
 #endif  // SRC_PANDARGENERAL_INTERNAL_H_
